% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfnDescstat.R
\name{bsfnDescstat}
\alias{bsfnDescstat}
\title{Calculate descriptive statistics of simulated PK parameters}
\usage{
bsfnDescstat(bsfnPkparamData)
}
\arguments{
\item{bsfnPkparamData}{data frame generated by bsfnPkparam function}
}
\value{
The descriptive statistics of pharmacokinetic parameters
}
\description{
\code{bsfnDescstat} will calculate descriptive statistics of simulated PK parameters
}
\examples{
bsfnDescstat(bsfnPkparam(20,500))
bsfnDescstat(bsfnPkparamMulti(20,500))
bsfnDescExample <- cbind(bsfnDescstat(bsfnPkparam(20,500)), 
                         bsfnDescstat(bsfnPkparam(50,500))[,2])
colnames(bsfnDescExample)[2:3] <- c('20 kg', '50 kg')      
bsfnDescExample
}
\seealso{
\url{https://asancpt.github.io/bsfnsim}
}
