% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfnOverdose.R
\name{bsfnOverdose}
\alias{bsfnOverdose}
\title{Calculate a duration of plasma busulfan concentration over specified toxic limits}
\usage{
bsfnOverdose(bsfnConcTimeData)
}
\arguments{
\item{bsfnConcTimeData}{data frame containing concentration-time data}
}
\value{
descriptive statistics of duration of toxic concentrations
}
\description{
\code{bsfnOverdose} calculates a time duration of plasma busulfan concentration over specified toxic limits (40 mg/L or 80 mg/L)
}
\examples{
bsfnOverdose(bsfnConcTime(Weight = 20, Dose = 200, N = 20))
bsfnOverdose(bsfnConcTimeMulti(Weight = 20, Dose = 200, N = 20, Tau = 8, Repeat = 4))
}
\seealso{
\url{https://asan.shinyapps.io/bsfn/}
}
