% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfnPkparam.R
\name{bsfnPkparam}
\alias{bsfnPkparam}
\title{Create a dataset for simulation of single dose of busulfan}
\usage{
bsfnPkparam(Weight, Dose, N = 20)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of single busulfan (mg)}

\item{N}{The number of simulated subjects}
}
\value{
The dataset of pharmacokinetic parameters of subjects after single busulfan dose following multivariate normal
}
\description{
\code{bsfnPkparam} will create a dataset for simulation of single dose of busulfan
}
\examples{
bsfnPkparam(Weight = 20, Dose = 200, N = 20)
bsfnPkparam(20,500)
}
\seealso{
\url{https://asancpt.github.io/bsfnsim}
}
