% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfnPkparam.R
\name{bsfnPkparamMulti}
\alias{bsfnPkparamMulti}
\title{Create a dataset for simulation of multiple dose of busulfan}
\usage{
bsfnPkparamMulti(Weight, Dose, N = 20, Tau = 8)
}
\arguments{
\item{Weight}{Body weight (kg)}

\item{Dose}{Dose of multiple busulfan (mg)}

\item{N}{The number of simulated subjects}

\item{Tau}{The interval of multiple dosing (hour)}
}
\value{
The dataset of pharmacokinetic parameters of subjects after multiple busulfan dose following multivariate normal
}
\description{
\code{bsfnPkparamMulti} will create a dataset for simulation of multiple dose of busulfan
}
\examples{
bsfnPkparamMulti(Weight = 20, Dose = 200, N = 20, Tau = 8)
bsfnPkparamMulti(20,500)
}
\seealso{
\url{https://asancpt.github.io/bsfnsim}
}
