% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfnPlot.R
\name{bsfnPlot}
\alias{bsfnPlot}
\title{Create concentration-time curve after single dose of busulfan}
\usage{
bsfnPlot(bsfnConcTimeData, log = FALSE)
}
\arguments{
\item{bsfnConcTimeData}{data frame of concentration-time dataset having column names Subject, Time, and Conc (case-sensitive)}

\item{log}{y axis log}
}
\value{
The concentration-time curve
}
\description{
\code{bsfnPlot} will create concentration-time curve after single dose of busulfan
}
\examples{
bsfnPlot(bsfnConcTime(Weight = 20, Dose = 200, N = 20))
}
\seealso{
\url{https://asancpt.github.io/bsfnsim}
}
