% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfnShiny.R
\name{bsfnShiny}
\alias{bsfnShiny}
\title{Run shiny app to interactively simulate plasma busulfan concentration.}
\usage{
bsfnShiny()
}
\description{
\code{bsfnShiny} runs an internal shiny app \code{Busulfan Concentration Predictor} in order to interactively simulate plasma busulfan concentration.
}
\seealso{
\url{https://asan.shinyapps.io/bsfn/}
}
